Refactor container names so it is clear they are complete pages that hold multiple components.
